refactor(factories): consolidate factories to composition/ layer
Move all factories from infrastructure/factories/ to composition/factories/
following the architectural requirement that factories belong in composition root.

Changes:
- Create composition/factories/storage_factory.py (merged storage.py + storage_factory.py)
- Create composition/factories/clients.py (create_redis_client, get_aws_credentials)
- Create composition/factories/data_sources.py (DataSourceFactory)
- Update composition/factories/__init__.py with new exports
- Update imports in base_services_factory, bootstrap, data_source_registry, uniprot_protein
- Add deprecation re-exports in infrastructure/factories/ for backwards compatibility
- Update test files to patch new module paths

All factory imports now go through composition/factories/, with deprecation warnings
for old infrastructure/factories/ paths to allow gradual migration.
