+++ src/bioetl/composition/factories/data_source_registry.py
@@ -10,9 +10,9 @@ from __future__ import annotations
 from typing import TYPE_CHECKING, Protocol
 
 from bioetl.application.core.filtered_data_source import FilteredDataSource
+from bioetl.composition.factories.data_sources import DataSourceFactory
 from bioetl.composition.factories.http_client_factory import HttpClientFactory
 from bioetl.infrastructure.adapters.input.csv_filter_reader import CsvFilterReader
-from bioetl.infrastructure.factories.data_sources import DataSourceFactory
 
 if TYPE_CHECKING:
     from bioetl.domain.filter_config import InputFilterConfig
